Add feature-comparison against systemd
Generated from the systemd man pages from the systemd repo at v244.

This is the result of running the included script.  It will require fixing
up manually.  I'm commiting this unmodified version so the script can be
rerun with future versions of systemd and a 3-way merge can be done to
keep the table up-to-date.